+++ examples/array_write_read.rs
@@ -45,7 +45,7 @@ fn array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     )
     // .bytes_to_bytes_codecs(vec![]) // uncompressed
     .dimension_names(Some(vec!["y".into(), "x".into()]))
-    .storage_transformers(vec![])
+    // .storage_transformers(vec![].into())
     .build(store.clone(), array_path)?;
 
     // Write array metadata to store